+++ scanpipe/views.py
@@ -1286,6 +1286,9 @@ class DiscoveredPackageListView(
         "resources",
     ]
 
+    def get_queryset(self):
+        return super().get_queryset().order_by_purl()
+
     def get_context_data(self, **kwargs):
         context = super().get_context_data(**kwargs)
         context["display_compliance_alert"] = scanpipe_app.policies_enabled